+++ src/entities.rs
@@ -7,3 +7,5 @@ pub(crate) mod variables;
 pub(crate) mod programming_languages;
 pub(crate) mod environment;
 pub(crate) mod auto_version;
+pub(crate) mod requirements;
+pub(crate) mod path_type;