+++ src/main.rs
@@ -1,4 +1,4 @@
-#![feature(let_chains, if_let_guard, once_wait, string_from_utf8_lossy_owned)]
+#![feature(let_chains, if_let_guard, once_wait, string_from_utf8_lossy_owned, str_as_str)]
 #![deny(warnings, clippy::todo, clippy::unimplemented)]
 
 #[cfg(feature = "tests")]
@@ -66,7 +66,7 @@ fn main() {
     if err.contains("called `Result::unwrap()` on an `Err` value: ") {
       eprintln!("{}", err.split("called `Result::unwrap()` on an `Err` value: ").last().unwrap());
     } else {
-      eprintln!("{}", err.split('\n').last().unwrap());
+      eprintln!("{}", err.split('\n').next_back().unwrap());
     }
     std::process::exit(1);
   }));