+++ src/tui/edit.rs
@@ -1,18 +1,102 @@
 use colored::Colorize;
+use std::collections::HashSet;
 use std::path::PathBuf;
 
 use crate::actions::{Action, DescribedAction};
-use crate::configs::DeployerGlobalConfig;
+use crate::configs::{DeployerGlobalConfig, DeployerProjectOptions};
 use crate::entities::custom_command::{CustomCommand, specify_bash_c};
+use crate::entities::path_type::PathType;
 use crate::entities::programming_languages::{ProgrammingLanguage, specify_programming_languages};
+use crate::entities::requirements::Requirement;
 use crate::entities::targets::{TargetDescription, OsVariant, OsVersionSpecification};
 use crate::entities::traits::{Edit, EditExtended};
 use crate::entities::variables::{Variable, VarValue};
 use crate::hmap;
 use crate::i18n;
 use crate::pipelines::DescribedPipeline;
-use crate::tui::add::{collect_af_inplacement, collect_artifact};
+use crate::tui::add::{collect_af_inplacement, collect_path, collect_artifact};
 use crate::utils::tags_custom_type;
+
+impl DeployerProjectOptions {
+  pub(crate) fn edit_project_from_prompt(&mut self, globals: &mut DeployerGlobalConfig) -> anyhow::Result<()> {
+    let actions = vec![
+      i18n::EDIT_PROJECT_PIPELINES,
+      i18n::EDIT_DEFAULT,
+      i18n::EDIT_PROJECT_NAME,
+      i18n::EDIT_PROJECT_REASSIGN,
+      i18n::EDIT_CACHE,
+      i18n::EDIT_PLS,
+      i18n::EDIT_TARGETS,
+      i18n::EDIT_DEPL_TOOLKIT,
+      i18n::EDIT_PROJECT_VARS,
+      i18n::EDIT_ARTIFACTS,
+      i18n::EDIT_AF_INPLACE,
+    ];
+    
+    while let Some(action) = inquire::Select::new(
+      &format!("{} {}:", i18n::EDIT_ACTION_PROMPT, i18n::HIT_ESC),
+      actions.clone(),
+    ).prompt_skippable()? {
+      match action {
+        i18n::EDIT_PROJECT_NAME => self.project_name = inquire::Text::new(i18n::PROJECT_NAME).prompt()?,
+        i18n::EDIT_DEFAULT => self.select_default_pipeline()?,
+        i18n::EDIT_CACHE => self.cache_files.edit_from_prompt()?,
+        i18n::EDIT_PLS => self.langs.edit_from_prompt()?,
+        i18n::EDIT_TARGETS => self.targets.edit_from_prompt()?,
+        i18n::EDIT_DEPL_TOOLKIT => self.deploy_toolkit = inquire::Text::new(
+          &format!("{} {}:", i18n::DEPL_TOOLKIT, i18n::OR_HIT_ESC)
+        ).prompt_skippable()?,
+        i18n::EDIT_PROJECT_VARS => self.variables.edit_from_prompt()?,
+        i18n::EDIT_ARTIFACTS => {
+          let mut path_type = PathType::Relative;
+          self.artifacts.edit_from_prompt(&mut path_type)?
+        },
+        i18n::EDIT_AF_INPLACE => self.inplace_artifacts_into_project_root.edit_from_prompt(&mut self.artifacts)?,
+        i18n::EDIT_PROJECT_PIPELINES => self.pipelines.edit_from_prompt(globals)?,
+        i18n::EDIT_PROJECT_REASSIGN => for pipeline in &mut self.pipelines {
+          for action in &mut pipeline.actions {
+            *action = action.prompt_setup_for_project(&self.langs, &self.deploy_toolkit, &self.targets, &self.variables, &self.artifacts)?;
+          }
+        },
+        _ => {},
+      }
+    }
+    
+    Ok(())
+  }
+  
+  pub(crate) fn select_default_pipeline(&mut self) -> anyhow::Result<()> {
+    match self.pipelines.len() {
+      0 => {
+        println!("{}", i18n::PROJECT_NO_PIPELINES);
+      },
+      1 => {
+        let pipeline = self.pipelines.first_mut().unwrap();
+        pipeline.default = Some(true);
+      },
+      _ => {
+        let mut cmap = hmap!();
+        let mut cs = vec![];
+        
+        self.pipelines.iter_mut().for_each(|c| {
+          let s = format!("{} `{}`", i18n::PIPELINE, c.title);
+          
+          cmap.insert(s.clone(), c);
+          cs.push(s);
+        });
+        
+        if let Some(pipe) = inquire::Select::new(&format!("{} {}:", i18n::EDIT_DEFAULT_PROMPT, i18n::HIT_ESC), cs).prompt_skippable()? {
+          for (key, val) in cmap.iter_mut() {
+            if key.as_str().eq(pipe.as_str()) { val.default = Some(true); }
+            else { val.default = Some(false); }
+          }
+        }
+      },
+    }
+    
+    Ok(())
+  }
+}
 
 impl EditExtended<DeployerGlobalConfig> for Vec<DescribedAction> {
   fn edit_from_prompt(&mut self, opts: &mut DeployerGlobalConfig) -> anyhow::Result<()> {
@@ -143,6 +227,7 @@ impl DescribedAction {
       i18n::EDIT_TITLE,
       i18n::EDIT_DESC,
       i18n::EDIT_TAGS,
+      i18n::EDIT_REQS,
     ]);
     
     while let Some(action) = inquire::Select::new(
@@ -210,6 +295,7 @@ impl DescribedAction {
           }
         },
         i18n::EDIT_PATCH if let Action::Patch(patch) = &mut self.action => { patch.edit_from_prompt()?; },
+        i18n::EDIT_REQS => self.requirements.edit_from_prompt()?,
         _ => {},
       }
     }
@@ -361,13 +447,158 @@ impl EditExtended<DeployerGlobalConfig> for Vec<DescribedPipeline> {
   }
 }
 
-impl Edit for Vec<PathBuf> {
+impl Requirement {
+  pub(crate) fn edit_requirement_from_prompt(&mut self) -> anyhow::Result<()> {
+    match self {
+      Self::Exists(path) => { *path = PathBuf::from(
+        inquire::Text::new(i18n::ABSOLUTE_PATH).with_default(path.to_str().unwrap()).prompt()?
+      )},
+      Self::ExistsAny(paths) => {
+        let mut path_type = PathType::Absolute;
+        paths.edit_from_prompt(&mut path_type)?;
+      },
+      Self::CheckSuccess(check_action) => check_action.edit_check_from_prompt()?,
+    }
+    Ok(())
+  }
+}
+
+impl Edit for Option<Vec<Requirement>> {
+  fn edit_from_prompt(&mut self) -> anyhow::Result<()> {
+    let mut reqs = match self {
+      None => vec![],
+      Some(reqs) => reqs.to_owned(),
+    };
+    reqs.edit_from_prompt()?;
+    if reqs.is_empty() { *self = None; } else { *self = Some(reqs); }
+    
+    Ok(())
+  }
+}
+
+impl Edit for Vec<Requirement> {
   fn edit_from_prompt(&mut self) -> anyhow::Result<()> {
     loop {
       let mut cmap = hmap!();
       let mut cs = vec![];
       
       self.iter_mut().for_each(|c| {
+        let s = format!("{} {}", i18n::REQUIREMENT, c);
+        cmap.insert(s.clone(), c);
+        cs.push(s);
+      });
+      
+      cs.extend_from_slice(&[i18n::ADD.to_string(), i18n::REMOVE.to_string()]);
+      
+      if let Some(action) = inquire::Select::new(&format!("{} {}:", i18n::EDIT_ACTION_PROMPT, i18n::HIT_ESC), cs).prompt_skippable()? {
+        match action.as_str() {
+          i18n::ADD => self.add_item()?,
+          i18n::REMOVE => self.remove_item()?,
+          s if cmap.contains_key(s) => cmap.get_mut(s).unwrap().edit_requirement_from_prompt()?,
+          _ => {},
+        }
+      } else { break }
+    }
+    
+    Ok(())
+  }
+  
+  fn add_item(&mut self) -> anyhow::Result<()> {
+    self.push(Requirement::new_from_prompt()?);
+    Ok(())
+  }
+  
+  fn remove_item(&mut self) -> anyhow::Result<()> {
+    let mut cmap = hmap!();
+    let mut cs = vec![];
+    
+    self.iter().for_each(|c| {
+      let s = format!("{} {}", i18n::REQUIREMENT, c);
+      cmap.insert(s.clone(), c);
+      cs.push(s);
+    });
+    
+    let selected = inquire::Select::new(i18n::VALUE_TO_REMOVE, cs.clone()).prompt()?;
+    
+    let mut reqs = vec![];
+    for key in cs {
+      if key.as_str().eq(selected.as_str()) { continue }
+      reqs.push((*cmap.get(&key).unwrap()).clone());
+    }
+    
+    *self = reqs;
+    Ok(())
+  }
+}
+
+impl EditExtended<PathType> for Vec<PathBuf> {
+  fn edit_from_prompt(&mut self, opts: &mut PathType) -> anyhow::Result<()> {
+    loop {
+      let mut cmap = hmap!();
+      let mut cs = vec![];
+      
+      self.iter().for_each(|c| {
+        let s = format!("{} `{}`", i18n::ENTITY, c.to_string_lossy());
+        
+        cmap.insert(s.clone(), c);
+        cs.push(s);
+      });
+      
+      cs.extend_from_slice(&[i18n::ADD.to_string(), i18n::REMOVE.to_string()]);
+      
+      if let Some(action) = inquire::Select::new(&format!("{} {}:", i18n::EDIT_ACTION_PROMPT, i18n::HIT_ESC), cs).prompt_skippable()? {
+        match action.as_str() {
+          i18n::ADD => self.add_item(opts)?,
+          i18n::REMOVE => self.remove_item(opts)?,
+          _ => {},
+        }
+      } else { break }
+    }
+    
+    Ok(())
+  }
+  
+  fn reorder(&mut self, _: &mut PathType) -> anyhow::Result<()> { Ok(()) }
+  
+  fn add_item(&mut self, opts: &mut PathType) -> anyhow::Result<()> {
+    match opts {
+      PathType::Absolute => self.push(collect_path()?),
+      PathType::Relative => self.push(collect_artifact()?),
+    }
+    Ok(())
+  }
+  
+  fn remove_item(&mut self, _: &mut PathType) -> anyhow::Result<()> {
+    let mut cmap = hmap!();
+    let mut cs = vec![];
+    
+    self.iter().for_each(|c| {
+      let s = format!("{} `{}`", i18n::ENTITY, c.to_string_lossy());
+      
+      cmap.insert(s.clone(), c);
+      cs.push(s);
+    });
+    
+    let selected = inquire::Select::new(i18n::VALUE_TO_REMOVE, cs.clone()).prompt()?;
+    
+    let mut commands = vec![];
+    for key in cs {
+      if key.as_str().eq(selected.as_str()) { continue }
+      commands.push((*cmap.get(&key).unwrap()).clone());
+    }
+    
+    *self = commands;
+    Ok(())
+  }
+}
+
+impl Edit for HashSet<PathBuf> {
+  fn edit_from_prompt(&mut self) -> anyhow::Result<()> {
+    loop {
+      let mut cmap = hmap!();
+      let mut cs = vec![];
+      
+      self.iter().for_each(|c| {
         let s = format!("{} `{}`", i18n::ENTITY, c.to_string_lossy());
         
         cmap.insert(s.clone(), c);
@@ -391,7 +622,7 @@ impl Edit for Vec<PathBuf> {
   fn reorder(&mut self) -> anyhow::Result<()> { Ok(()) }
   
   fn add_item(&mut self) -> anyhow::Result<()> {
-    self.push(collect_artifact()?);
+    self.insert(collect_artifact()?);
     Ok(())
   }
   
@@ -408,10 +639,10 @@ impl Edit for Vec<PathBuf> {
     
     let selected = inquire::Select::new(i18n::VALUE_TO_REMOVE, cs.clone()).prompt()?;
     
-    let mut commands = vec![];
+    let mut commands = HashSet::new();
     for key in cs {
       if key.as_str().eq(selected.as_str()) { continue }
-      commands.push((*cmap.get(&key).unwrap()).clone());
+      commands.insert((*cmap.get(&key).unwrap()).clone());
     }
     
     *self = commands;
@@ -555,7 +786,11 @@ impl Variable {
       match action {
         i18n::EDIT_TITLE => self.title = inquire::Text::new(i18n::VAR_TITLE).prompt()?,
         i18n::EDIT_VAR_SECRET => self.is_secret = inquire::Confirm::new(i18n::VAR_IS_SECRET).with_default(false).prompt()?,
-        i18n::EDIT_VALUE => self.value = VarValue::Plain(inquire::Text::new(i18n::VAR_CONTENT).prompt()?),
+        i18n::EDIT_VALUE => match &mut self.value {
+          VarValue::Plain(plain) => *plain = inquire::Text::new(i18n::VAR_PLAIN_CONTENT).with_default(plain).prompt()?,
+          VarValue::FromEnvFile(_) => self.value = Variable::new_env_from_prompt()?,
+          VarValue::FromHCVaultKv2(_) => self.value = Variable::new_kv2_from_prompt()?,
+        },
         _ => {},
       }
     }